+++ conda/recipes/libcudf/meta.yaml
@@ -60,7 +60,7 @@ test:
     - test -f $PREFIX/include/cudf/filling.hpp
     - test -f $PREFIX/include/cudf/functions.h
     - test -f $PREFIX/include/cudf/gis.hpp
-    - test -f $PREFIX/include/cudf/groupby.hpp
+    - test -f $PREFIX/include/cudf/legacy/groupby.hpp
     - test -f $PREFIX/include/cudf/io_functions.h
     - test -f $PREFIX/include/cudf/io_functions.hpp
     - test -f $PREFIX/include/cudf/io_readers.hpp
