+++ cpp/src/groupby/common/legacy/utils.hpp
@@ -21,7 +21,7 @@
 #include <cudf/cudf.h>
 #include <bitmask/legacy/bit_mask.cuh>
 #include <cudf/legacy/copying.hpp>
-#include <cudf/groupby.hpp>
+#include <cudf/legacy/groupby.hpp>
 #include <cudf/legacy/bitmask.hpp>
 #include <cudf/legacy/table.hpp>
 #include <hash/concurrent_unordered_map.cuh>
