+++ cpp/src/quantiles/group_quantiles.cu
@@ -18,7 +18,7 @@
 
 #include <utilities/cuda_utils.hpp>
 #include <quantiles/quantiles_util.hpp>
-#include <groupby/sort/sort_helper.hpp>
+#include <groupby/sort/legacy/sort_helper.hpp>
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
 
 #include <cudf/cudf.h>
