+++ cpp/tests/groupby/sort/ignore_null_keys/legacy/single_column_avg.cu
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-#include "../single_column_groupby_test.cuh"
-#include "../../common/type_info.hpp"
-#include <cudf/groupby.hpp>
+#include "../../legacy/single_column_groupby_test.cuh"
+#include "../../../common/legacy/type_info.hpp"
+#include <cudf/legacy/groupby.hpp>
 #include <cudf/legacy/table.hpp>
 #include <cudf/utilities/legacy/type_dispatcher.hpp>
 #include <tests/utilities/legacy/column_wrapper.cuh>
