+++ cpp/tests/groupby/without_agg/legacy/groupby_wo_agg_test.cu
@@ -26,7 +26,7 @@
 #include <gmock/gmock.h>
 
 #include <cudf/cudf.h>
-#include <cudf/groupby.hpp>
+#include <cudf/legacy/groupby.hpp>
 #include <cudf/legacy/table.hpp>
 
 #include <utilities/cudf_utils.h>
