+++ python/cudf/cudf/_lib/includes/groupby/common.pxd
@@ -8,7 +8,7 @@
 from cudf._lib.cudf import *
 from cudf._lib.cudf cimport *
 
-cdef extern from "cudf/groupby.hpp" namespace "cudf::groupby" nogil:
+cdef extern from "cudf/legacy/groupby.hpp" namespace "cudf::groupby" nogil:
     cdef cppclass Options:
         Options(bool _ignore_null_keys) except +
         Options() except +
