+++ python/cudf/cudf/_lib/includes/groupby/hash.pxd
@@ -13,7 +13,7 @@ from cudf._lib.cudf cimport *
 
 cimport cudf._lib.includes.groupby.common as groupby_common
 
-cdef extern from "cudf/groupby.hpp" namespace "cudf::groupby" nogil:
+cdef extern from "cudf/legacy/groupby.hpp" namespace "cudf::groupby" nogil:
     ctypedef enum operators:
         SUM,
         MIN,
@@ -23,7 +23,7 @@ cdef extern from "cudf/groupby.hpp" namespace "cudf::groupby" nogil:
         MEDIAN,
         QUANTILE
 
-cdef extern from "cudf/groupby.hpp" namespace "cudf::groupby::hash" nogil:
+cdef extern from "cudf/legacy/groupby.hpp" namespace "cudf::groupby::hash" nogil:
     cdef cppclass Options(groupby_common.Options):
         Options(bool _ignore_null_keys) except +
