+++ python/cudf/cudf/_lib/includes/groupby/sort.pxd
@@ -10,7 +10,7 @@ from libcpp.utility cimport pair
 from cudf._lib.cudf import *
 from cudf._lib.cudf cimport *
 
-cdef extern from "cudf/groupby.hpp" nogil:
+cdef extern from "cudf/legacy/groupby.hpp" nogil:
     cdef pair[cudf_table, gdf_column] gdf_group_by_without_aggregations(
         const cudf_table  cols,
         gdf_size_type num_key_cols,
